merge

+++ src/controllers/product.controller.ts
@@ -2,15 +2,17 @@ import { Response, Request } from 'express'
 import ProductModel from '../models/product.model'
 import path from 'path'
 import fs from 'fs'
-import fs_async from 'fs/promises';
+import fs_async from 'fs/promises'
 import { IRequest } from 'types'
 import CategoryModel from '../models/category.model'
+import { Product } from 'shared/types/Product'
 export const getProducts = async (req: IRequest, res: Response) => {
 	try {
 		const products = await ProductModel.find()
 			.select('-__v')
 			.populate('parentCategory')
 			.populate('childCategory')
+
 		res.json(products)
 	} catch (err) {
 		res.status(500).json({ error: 'Eroare la obținerea produselor' })
@@ -20,6 +22,7 @@ export const getProducts = async (req: IRequest, res: Response) => {
 export const getProduct = async (req: IRequest, res: Response) => {
 	try {
 		const product = await ProductModel.findById(req.params.id)
+
 		if (product) {
 			res.json(product)
 		} else {
@@ -42,9 +45,7 @@ export const createProduct = async (req: Request, res: Response) => {
 			: null
 		console.log('parentCategoryDoc', parentCategoryDoc)
 		console.log('childCategoryDoc', childCategoryDoc)
-		const imageUrl = `/images/${
-			file.filename
-		}`
+		const imageUrl = `/images/${file.filename}`
 
 		const newProduct = new ProductModel({
 			name,
@@ -70,16 +71,12 @@ export const updateProduct = async (req: Request, res: Response) => {
 		const { name, description, price, quantityInGrams, parentId, childId } =
 			req.body
 		const file = req.file as Express.Multer.File
-		console.log(file)
-		console.log(req.body)
+
 		const product = await ProductModel.findById(req.params.id)
 		const parentCategoryDoc = await CategoryModel.findById(parentId)
 		const childCategoryDoc = await CategoryModel.findById(childId)
-		console.log('parentCategoryDoc', parentCategoryDoc)
-		console.log('childCategoryDoc', childCategoryDoc)
-		const imageUrl = file
-			? `/images/${file.filename}`
-			: product?.imageUrl
+
+		const imageUrl = file ? `/images/${file.filename}` : product?.imageUrl
 		const updatedProduct = await ProductModel.findByIdAndUpdate(
 			req.params.id,
 			{
@@ -114,11 +111,16 @@ export const updateProduct = async (req: Request, res: Response) => {
 
 export const deleteProduct = async (req: IRequest, res: Response) => {
 	try {
+		let imageUrl = ((await ProductModel.findById(req.params.id)) as Product)
+			.imageUrl
+		if (!imageUrl) {
+			return res.status(404).json({ message: 'Produsul nu a fost găsit' })
+		}
 		const deletedProduct = await ProductModel.findByIdAndDelete(
 			req.params.id
 		)
-		console.log(path.join(__dirname, `../../public/${deletedProduct?.imageUrl}`));
-		await fs_async.unlink(path.join(__dirname, `../../public/${deletedProduct?.imageUrl}`))
+
+		await fs_async.unlink(path.join(__dirname, `../../public/${imageUrl}`))
 
 		if (deletedProduct) {
 			res.status(200).json({ message: 'Produs șters cu succes' })
@@ -126,7 +128,10 @@ export const deleteProduct = async (req: IRequest, res: Response) => {
 			res.status(404).json({ message: 'Produsul nu a fost găsit' })
 		}
 	} catch (err) {
-		res.status(500).json({ error: 'Eroare la ștergerea produsului', msg: err})
+		res.status(500).json({
+			error: 'Eroare la ștergerea produsului',
+			msg: err,
+		})
 	}
 }
 
@@ -138,7 +143,10 @@ export const toggleProductVisibility = async (req: IRequest, res: Response) => {
 				req.params.id,
 				{ isVisible: !product.isVisible },
 				{ new: true }
-			).select('-__v')
+			)
+				.select('-__v')
+				.populate('parentCategory')
+				.populate('childCategory')
 			res.status(200).json(updatedProduct)
 		} else {
 			res.status(404).json({ message: 'Produsul nu a fost găsit' })
